pipeline import

+++ src/main/java/cc/altius/FASP/model/pipeline/QatTempProgramPlanningUnit.java
@@ -19,7 +19,7 @@ public class QatTempProgramPlanningUnit extends BaseModel implements Serializabl
     private String programPlanningUnitId;
     private SimpleObject program;
     private String planningUnitId;
-    private int multiplier;
+    private double multiplier;
     private int reorderFrequencyInMonths;
     private int minMonthsOfStock;
     private int productCategoryId;
@@ -34,7 +34,7 @@ public class QatTempProgramPlanningUnit extends BaseModel implements Serializabl
     public QatTempProgramPlanningUnit() {
     }
 
-    public QatTempProgramPlanningUnit(String programPlanningUnitId, SimpleObject program, String planningUnitId,int multiplier, int reorderFrequencyInMonths, int minMonthsOfStock, int productCategoryId, String pipelineProductName, String pipelineProductCategoryName, double localProcurmentLeadTime, int shelfLife, double catalogPrice, int monthsInPastForAmc, int monthsInFutureForAmc) {
+    public QatTempProgramPlanningUnit(String programPlanningUnitId, SimpleObject program, String planningUnitId,double multiplier, int reorderFrequencyInMonths, int minMonthsOfStock, int productCategoryId, String pipelineProductName, String pipelineProductCategoryName, double localProcurmentLeadTime, int shelfLife, double catalogPrice, int monthsInPastForAmc, int monthsInFutureForAmc) {
         this.programPlanningUnitId = programPlanningUnitId;
         this.program = program;
         this.planningUnitId = planningUnitId;
@@ -155,11 +155,11 @@ public class QatTempProgramPlanningUnit extends BaseModel implements Serializabl
         this.monthsInFutureForAmc = monthsInFutureForAmc;
     }
 
- public int getMultiplier() {
+ public double getMultiplier() {
         return multiplier;
     }
 
-    public void setMultiplier(int multiplier) {
+    public void setMultiplier(double multiplier) {
         this.multiplier = multiplier;
     }
